support starts_at, ends_at and refactor image routes
